Keep QIcon::fromTheme in main thread
Summary:
QIcon::fromTheme is not thread safe, we can't have it done in the future
watcher.
We can solve this by returning null in the watcher, and doing fromTheme
in the main thread
when the watcher finishes if we don't have an icon.

Test Plan: Current unit tests still pass

Reviewers: #plasma, graesslin

Reviewed By: #plasma, graesslin

Subscribers: graesslin, plasma-devel, #frameworks

Tags: #frameworks, #plasma

Differential Revision: https://phabricator.kde.org/D5889