[client] Drop the checks for platformName being "wayland"
Summary:
There are also QPA's which are "wayland" but have a different name. An
example is flatpak. Any code using KWayland::Client is basically
disabled on this platform due to the incorrect checks for the platform
name.

If we run on a platform which provides wl_display through the native
interface we should be able to assume that it truly is Wayland and that
it speaks wayland. Similar for all the other checks. The only
problematic code might be the check in Surface, but on the other hand we
can assume that code using Surface::fromWindow has also created a
ConnectionThread and thus knows whether it's Wayland or not.

This fixes the Plasma integration for flatpak.

Test Plan: Flatpak windows have decorations in Plasma now.

Reviewers: #frameworks, #kwin, #flatpak

Subscribers: plasma-devel

Tags: #plasma_on_wayland, #frameworks

Differential Revision: https://phabricator.kde.org/D8250