Fix naming inconsistency
Summary:
Everything was meant to be ServerSideDecorationPalette, the client and
server managers are named that.

The interface that attaches to the window was accidentally called
ServerDecorationPalette on the client side. (with the word "Side" missing)

Not an API break as current code is not yet released.

Test Plan:
Grepped for any others being out
Test didn't need changes as it used auto.

Reviewers: #plasma

Subscribers: plasma-devel, #frameworks

Tags: #plasma, #frameworks

Differential Revision: https://phabricator.kde.org/D9656