Fix windowmanagement test
We are comparing pixmaps that seem to be using different formats and
it's failing because of it, although it's not the intent of this test to
make sure that the format is exactly the same but actually that the
images are the same.
This patch changes them to offer the same image so they can be properly
compared.