Install Client headers in a dirs hierarchy matching the C++ namespaces
Headers should be in <prefix>/include/KF*/ModuleName/NameSpace, so
KF*/KWayland/KWayland/Client ...etc.

This is so that "#include <KWayland/Client/foo.h>" continues to work
with KF6. Thanks to Volker for explaining that to me :)

This auto-magically worked with KF5 because <prefix>/include/KF*/ was
always added to the targets interface include dirs by ECM (for
compatibility with KDE4?).

See:
https://mail.kde.org/pipermail/kde-frameworks-devel/2022-January/120608.html

Note the Server part wasn't changed by this commit because it's deprecated
as it was moved to Plasma to accommodate the KWin release cycle.