Flush display also on awake
aboutToBlock is only emitted when events are processes with
QEventLoop::WaitForMoreEvents. A event loop iteration emits
'aboutToBlock', processes/waits for events and emits 'awake'.
Notably qWait which is used by QTRY_ macros processes events
manually without this flag. In such a case only 'awake' is emitted
at the start of an iteration.
Flushing on awake enables us to use KWayland in tests together with
QTRY_ macros. This also matches the QtWayland implementation.